Add authorization based on tenant

+++ src/tanbyiot.client/src/api/RulesApi.ts
@@ -1,15 +1,15 @@
 ﻿import {
     Rule,
     RuleAction,
-    RuleActionType,
     RuleCondition,
-    RuleConditionType,
     RuleDetail, UpdateRuleActionArgs, UpdateRuleArgs,
     UpdateRuleConditionArgs
 } from "./types.t.ts";
 import {getApi} from "./Api.ts";
 import moment from "moment";
 import { useAuth0 } from "@auth0/auth0-react";
+import {useUser} from "./UsersApi.ts";
+import {RuleActionType, RuleConditionType} from "./enums.ts";
 
 function transformRuleFromServer(rule: Rule): Rule {
     return {
@@ -40,6 +40,7 @@ export function useRulesApi() {
         getAccessTokenSilently,
         isAuthenticated
     } = useAuth0();
+    const user = useUser();
 
     async function getAuthenticatedApi() {
         const token = await getAccessTokenSilently();
@@ -50,13 +51,13 @@ export function useRulesApi() {
         ready: isAuthenticated,
         getAllRules: async function(): Promise<Rule[]> {
             const api = await getAuthenticatedApi();
-            const response = await api.get<Rule[]>("api/rules");
+            const response = await api.get<Rule[]>(`/api/tenants/${user?.currentTenant.id}/rules`);
 
             return response.data.map(transformRuleFromServer);
         },
         addRule: async function(name: string): Promise<Rule> {
-            const api = getApi();
-            const response = await api.post<Rule>(`/api/rules`, {
+            const api = await getAuthenticatedApi();
+            const response = await api.post<Rule>(`/api/tenants/${user?.currentTenant.id}/rules`, {
                 name
             });
 
@@ -64,7 +65,7 @@ export function useRulesApi() {
         },
         getRule: async function(id: string): Promise<RuleDetail> {
             const api = await getAuthenticatedApi();
-            const response = await api.get<RuleDetail>(`/api/rules/${id}`);
+            const response = await api.get<RuleDetail>(`/api/tenants/${user?.currentTenant.id}/rules/${id}`);
 
             return {
                 ...response.data,
@@ -75,46 +76,45 @@ export function useRulesApi() {
         },
         updateRule: async function(id: string, args: UpdateRuleArgs): Promise<Rule> {
             const api = await getAuthenticatedApi();
-            const response = await api.put<Rule>(`/api/rules/${id}`, args);
+            const response = await api.put<Rule>(`/api/tenants/${user?.currentTenant.id}/rules/${id}`, args);
 
             return transformRuleFromServer(response.data);
         },
         addRuleCondition: async function(ruleId: string, type: RuleConditionType): Promise<RuleCondition> {
             const api = await getAuthenticatedApi();
-            const response = await api.post<RuleCondition>(`/api/rules/${ruleId}/conditions`, {
+            const response = await api.post<RuleCondition>(`/api/tenants/${user?.currentTenant.id}/rules/${ruleId}/conditions`, {
                 type
             });
 
             return transformRuleConditionFromServer(response.data);
         },
-        updateRuleCondition: async function(ruleId: string, ruleConditionId: string, args: UpdateRuleConditionArgs): Promise<RuleCondition> {
+        updateRuleCondition: async function(ruleId: string, args: UpdateRuleConditionArgs): Promise<RuleCondition> {
             const api = await getAuthenticatedApi();
-            const response = await api.put(`/api/rules/${ruleId}/conditions/${ruleConditionId}`, args);
+            const response = await api.put(`/api/tenants/${user?.currentTenant.id}/rules/${ruleId}/conditions`, args);
 
             return transformRuleConditionFromServer(response.data);
         },
         deleteRuleCondition: async function(ruleId: string, ruleConditionId: string) {
             const api = await getAuthenticatedApi();
-            await api.delete(`/api/rules/${ruleId}/conditions/${ruleConditionId}`);
+            await api.delete(`/api/tenants/${user?.currentTenant.id}/rules/${ruleId}/conditions/${ruleConditionId}`);
         },
         addRuleAction: async function(ruleId: string, type: RuleActionType): Promise<RuleAction> {
             const api = await getAuthenticatedApi();
-            const response = await api.post<RuleAction>(`/api/rules/${ruleId}/actions`, {
+            const response = await api.post<RuleAction>(`/api/tenants/${user?.currentTenant.id}/rules/${ruleId}/actions`, {
                 type
             });
 
             return transformRuleActionFromServer(response.data);
         },
         updateRuleAction: async function(ruleId: string, ruleActionId: string, args: UpdateRuleActionArgs): Promise<RuleAction> {
-            console.log(args);
             const api = await getAuthenticatedApi();
-            const response = await api.put(`/api/rules/${ruleId}/actions/${ruleActionId}`, args);
+            const response = await api.put(`/api/tenants/${user?.currentTenant.id}/rules/${ruleId}/actions/${ruleActionId}`, args);
 
             return transformRuleActionFromServer(response.data);
         },
         deleteRuleAction: async function(ruleId: string, ruleActionId: string) {
             const api = await getAuthenticatedApi();
-            await api.delete(`/api/rules/${ruleId}/actions/${ruleActionId}`);
+            await api.delete(`/api/tenants/${user?.currentTenant.id}/rules/${ruleId}/actions/${ruleActionId}`);
         }
     };
 }
